# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Flash Deals API — premium same-day deal platform for upscale venues
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
  code?: string;
}

export interface Pagination {
  total: number;
  limit: number;
  offset: number;
}

export type VenueCategory = (typeof VenueCategory)[keyof typeof VenueCategory];

export const VenueCategory = {
  restaurant: "restaurant",
  spa: "spa",
  bar: "bar",
  fitness: "fitness",
  experience: "experience",
} as const;

export type VenueNeighborhood =
  (typeof VenueNeighborhood)[keyof typeof VenueNeighborhood];

export const VenueNeighborhood = {
  westlands: "westlands",
  kilimani: "kilimani",
  cbd: "cbd",
  karen: "karen",
  langata: "langata",
  lavington: "lavington",
  kileleshwa: "kileleshwa",
  runda: "runda",
  muthaiga: "muthaiga",
  gigiri: "gigiri",
  upper_hill: "upper_hill",
  other: "other",
} as const;

export type VenueStatus = (typeof VenueStatus)[keyof typeof VenueStatus];

export const VenueStatus = {
  pending_approval: "pending_approval",
  approved: "approved",
  suspended: "suspended",
} as const;

export type DealStatus = (typeof DealStatus)[keyof typeof DealStatus];

export const DealStatus = {
  draft: "draft",
  live: "live",
  filling_fast: "filling_fast",
  sold_out: "sold_out",
  expired: "expired",
  cancelled: "cancelled",
} as const;

export type DealCategory = (typeof DealCategory)[keyof typeof DealCategory];

export const DealCategory = {
  lunch: "lunch",
  dinner: "dinner",
  brunch: "brunch",
  treatment: "treatment",
  class: "class",
  experience: "experience",
  drinks: "drinks",
  tasting: "tasting",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending_payment: "pending_payment",
  confirmed: "confirmed",
  checked_in: "checked_in",
  completed: "completed",
  cancelled: "cancelled",
  refunded: "refunded",
} as const;

export type LoyaltyTier = (typeof LoyaltyTier)[keyof typeof LoyaltyTier];

export const LoyaltyTier = {
  bronze: "bronze",
  silver: "silver",
  gold: "gold",
  platinum: "platinum",
} as const;

export interface Venue {
  id: number;
  name: string;
  slug: string;
  category: VenueCategory;
  neighborhood: VenueNeighborhood;
  address: string;
  latitude?: string | null;
  longitude?: string | null;
  description: string;
  coverImage?: string | null;
  images: string[];
  commissionRate: string;
  status: VenueStatus;
  averageRating?: string | null;
  totalRatings: number;
  totalBookings: number;
  fillRate?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateVenueRequest {
  name: string;
  slug: string;
  category: VenueCategory;
  neighborhood: VenueNeighborhood;
  address: string;
  description: string;
  whatsappNumber?: string;
  contactEmail?: string;
  tags?: string[];
}

export interface UpdateVenueRequest {
  name?: string;
  description?: string;
  coverImage?: string;
  images?: string[];
  whatsappNumber?: string;
  contactEmail?: string;
  websiteUrl?: string;
  tags?: string[];
}

export interface VenueListResponse {
  data: Venue[];
  pagination: Pagination;
}

export interface Deal {
  id: number;
  venueId: number;
  title: string;
  description: string;
  category: DealCategory;
  discountPercent: number;
  originalPrice: string;
  dealPrice: string;
  totalSlots: number;
  bookedSlots: number;
  availableSlots: number;
  status: DealStatus;
  startsAt: string;
  endsAt: string;
  imageUrl?: string | null;
  isStanding: boolean;
  viewCount: number;
  venue?: Venue | null;
  createdAt: string;
  updatedAt: string;
}

export interface VenueAnalytics {
  venueId: number;
  totalDeals: number;
  totalBookings: number;
  fillRate: number;
  averageRating: number;
  totalRevenue: number;
  topDealCategory?: string;
  bestDay?: string;
  bestHour?: number;
  recentDeals: Deal[];
}

export interface CreateDealRequest {
  venueId: number;
  title: string;
  description: string;
  category: DealCategory;
  discountPercent: number;
  originalPrice: string;
  dealPrice: string;
  totalSlots: number;
  startsAt: string;
  endsAt: string;
  imageUrl?: string;
  isStanding?: boolean;
  standingDaysOfWeek?: number[];
  autoActivateThreshold?: number;
}

export interface UpdateDealRequest {
  title?: string;
  description?: string;
  totalSlots?: number;
  startsAt?: string;
  endsAt?: string;
  imageUrl?: string;
}

export interface DealListResponse {
  data: Deal[];
  pagination: Pagination;
}

export interface Booking {
  id: number;
  dealId: number;
  userId: number;
  venueId: number;
  slots: number;
  totalAmount: string;
  commissionAmount: string;
  venueAmount: string;
  status: BookingStatus;
  mpesaRef?: string | null;
  confirmationCode: string;
  specialRequests?: string | null;
  isCorporate: boolean;
  deal?: Deal | null;
  venue?: Venue | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateBookingRequest {
  dealId: number;
  slots: number;
  specialRequests?: string;
  isCorporate?: boolean;
  corporateName?: string;
  phoneNumber: string;
}

export interface BookingListResponse {
  data: Booking[];
  pagination: Pagination;
}

export interface User {
  id: number;
  email: string;
  phone?: string | null;
  name: string;
  role: string;
  loyaltyTier: LoyaltyTier;
  loyaltyPoints: number;
  subscriptionCategories: string[];
  neighborhoodPref?: string | null;
  createdAt: string;
}

export interface UpdateUserRequest {
  name?: string;
  phone?: string;
  subscriptionCategories?: string[];
  neighborhoodPref?: string;
  latitude?: string;
  longitude?: string;
  pushToken?: string;
}

export interface Rating {
  id: number;
  bookingId: number;
  userId: number;
  venueId: number;
  dealId: number;
  score: number;
  comment?: string | null;
  response?: string | null;
  createdAt: string;
}

export interface CreateRatingRequest {
  bookingId: number;
  score: number;
  comment?: string;
}

export interface RatingListResponse {
  data: Rating[];
  pagination: Pagination;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type ListVenuesParams = {
  category?: VenueCategory;
  neighborhood?: VenueNeighborhood;
  limit?: number;
  offset?: number;
};

export type ListDealsParams = {
  category?: DealCategory;
  neighborhood?: VenueNeighborhood;
  venueId?: number;
  status?: DealStatus;
  limit?: number;
  offset?: number;
};

export type ListBookingsParams = {
  status?: BookingStatus;
  limit?: number;
  offset?: number;
};

export type ListVenueRatingsParams = {
  limit?: number;
};