# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nairobi Flash Deals API — premium same-day deal platform for upscale venues
 * OpenAPI spec version: 0.2.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Deal } from "./deal";
import type { Venue } from "./venue";

export interface Booking {
  id: number;
  dealId: number;
  userId: number;
  venueId: number;
  slots: number;
  totalAmount: string;
  commissionAmount: string;
  venueAmount: string;
  status: BookingStatus;
  mpesaRef?: string | null;
  confirmationCode: string;
  specialRequests?: string | null;
  isCorporate: boolean;
  deal?: Deal | null;
  venue?: Venue | null;
  createdAt: Date;
  updatedAt: Date;
}